+++ doc/CHANGES.rst
@@ -21,6 +21,8 @@ CHANGELOG
   decorator with alias in AnyBlok.
 * Fixed #60: Now DateTime plugins verify also the DateTime columns of the dependencies of the Model
 * Removed **Python 3.3** compatibility
+* Improved the NoResultFound Exception for query.one and query.dictone. Now the registry name of the model
+  isadded in the exception's message
 
 0.20.0 (2018-09-10)
 -------------------